refactor: separate core GUI and application layers
Split advertisement parsing and AirPods state management out of the device facade, and introduce a shared GUI context for window dependencies. Reorganize CMake targets so the core and GUI layers can be built and validated independently.
