+++ Source/Core/Settings.h
@@ -20,6 +20,7 @@
 
 #include <mutex>
 
+#include <QLocale>
 #include <QSettings>
 
 #include "../Helper.h"
@@ -29,6 +30,25 @@ namespace Core::Settings {
 enum class TrayIconBatteryBehavior : uint32_t { Disable, WhenLowBattery, Always };
 enum class TaskbarStatusBehavior : uint32_t { Disable, Text, Icon };
 enum class LoadResult : uint32_t { AbiIncompatible, NoAbiField, Successful };
+
+// Receives settings side effects so that this module doesn't need to know
+// about GUI or other application services.
+//
+class ApplyObserver
+{
+public:
+    virtual ~ApplyObserver() = default;
+
+    virtual void OnLanguageLocaleChanged(const QLocale &locale) = 0;
+    virtual void OnLowAudioLatencyChanged(bool enable) = 0;
+    virtual void OnAutomaticEarDetectionChanged(bool enable) = 0;
+    virtual void OnRssiMinChanged(int16_t rssiMin) = 0;
+    virtual void OnDeviceAddressChanged(uint64_t address) = 0;
+    virtual void OnTrayIconBatteryChanged(TrayIconBatteryBehavior behavior) = 0;
+    virtual void OnTaskbarBatteryChanged(TaskbarStatusBehavior behavior) = 0;
+};
+
+void SetApplyObserver(ApplyObserver *observer);
 
 // TODO: in [v1.0.0] [kFieldsAbiVersion = 2]
 //        - Rename `tray_icon_battery` to `battery_on_tray_icon`
