+++ Source/Gui/MainWindow.h
@@ -26,7 +26,7 @@
 #include <QMediaPlayer>
 #include <QPropertyAnimation>
 
-#include "../Utils.h"
+#include "Utils.h"
 #include "../Core/AirPods.h"
 #include "../Core/Update.h"
 #include "Base.h"
@@ -49,11 +49,6 @@ class MainWindow : public QDialog
 
 public:
     MainWindow(QWidget *parent = nullptr);
-
-    inline auto &GetApdMgr()
-    {
-        return _apdMgr;
-    }
 
     void UpdateState(const Core::AirPods::State &state);
     void Available();
@@ -88,7 +83,6 @@ private:
     Widget::Battery *_rightBattery = new Widget::Battery{this};
     Widget::Battery *_caseBattery = new Widget::Battery{this};
 
-    Core::AirPods::Manager _apdMgr;
     Core::Update::AsyncChecker _updateChecker{[this](auto &&...args) {
         VersionUpdateAvailableSafely(std::forward<decltype(args)>(args)...);
     }};
